+++ cli/CHANGELOG.md
@@ -4,6 +4,22 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.132.0](https://github.com/wundergraph/cosmo/compare/wgc@0.131.2...wgc@0.132.0) (2026-09-25)
+
+### Features
+
+* bump go version ([#3277](https://github.com/wundergraph/cosmo/issues/3277)) ([83d96f5](https://github.com/wundergraph/cosmo/commit/83d96f588b6a4b7f6c5def9801a2dd5a140bc75a)) (@Noroth)
+
+## [0.131.2](https://github.com/wundergraph/cosmo/compare/wgc@0.131.1...wgc@0.131.2) (2026-09-23)
+
+### Bug Fixes
+
+* increase timeout settings for client, build server, and routes ([#3274](https://github.com/wundergraph/cosmo/issues/3274)) ([1e77927](https://github.com/wundergraph/cosmo/commit/1e779279f0f1b677d164ee3deb55be16418ab415)) (@JivusAyrus)
+
+## [0.131.1](https://github.com/wundergraph/cosmo/compare/wgc@0.131.0...wgc@0.131.1) (2026-09-16)
+
+**Note:** Version bump only for package wgc
+
 # [0.131.0](https://github.com/wundergraph/cosmo/compare/wgc@0.130.4...wgc@0.131.0) (2026-09-14)
 
 ### Features