+++ router-plugin/CHANGELOG.md
@@ -4,6 +4,12 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.9.0](https://github.com/wundergraph/cosmo/compare/router-plugin@0.8.4...router-plugin@0.9.0) (2026-09-25)
+
+### Features
+
+* bump go version ([#3277](https://github.com/wundergraph/cosmo/issues/3277)) ([83d96f5](https://github.com/wundergraph/cosmo/commit/83d96f588b6a4b7f6c5def9801a2dd5a140bc75a)) (@Noroth)
+
 ## [0.8.4](https://github.com/wundergraph/cosmo/compare/router-plugin@0.8.3...router-plugin@0.8.4) (2026-09-09)
 
 **Note:** Version bump only for package router-plugin