+++ studio/CHANGELOG.md
@@ -4,6 +4,20 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.182.0](https://github.com/wundergraph/cosmo/compare/studio@0.181.2...studio@0.182.0) (2026-09-25)
+
+### Features
+
+* bump go version ([#3277](https://github.com/wundergraph/cosmo/issues/3277)) ([83d96f5](https://github.com/wundergraph/cosmo/commit/83d96f588b6a4b7f6c5def9801a2dd5a140bc75a)) (@Noroth)
+
+## [0.181.2](https://github.com/wundergraph/cosmo/compare/studio@0.181.1...studio@0.181.2) (2026-09-23)
+
+**Note:** Version bump only for package studio
+
+## [0.181.1](https://github.com/wundergraph/cosmo/compare/studio@0.181.0...studio@0.181.1) (2026-09-16)
+
+**Note:** Version bump only for package studio
+
 # [0.181.0](https://github.com/wundergraph/cosmo/compare/studio@0.180.5...studio@0.181.0) (2026-09-14)
 
 ### Features